Re-organised EventDetails task, state & reducer by splitting from EventsList
