+++ src/apps/events/client/Details/state.js
@@ -1,0 +1,9 @@
+export const TASK_GET_EVENT_DETAILS = 'TASK_GET_EVENT_DETAILS'
+
+export const ID = 'eventDetails'
+
+export const state2props = ({ ...state }) => {
+  return {
+    ...state[ID],
+  }
+}
