+++ src/apps/events/client/Details/tasks.js
@@ -1,0 +1,7 @@
+import { apiProxyAxios } from '../../../../client/components/Task/utils'
+import { transformResponseToEventDetails } from '../transformers'
+
+export const getEventDetails = (eventId) =>
+  apiProxyAxios
+    .get(`/api-proxy/v3/event/${eventId}`)
+    .then(({ data }) => transformResponseToEventDetails(data))
