+++ src/apps/events/client/reducer.js
@@ -2,7 +2,6 @@ import {
   EVENTS__LOADED,
   EVENTS__METADATA_LOADED,
   EVENTS__SELECTED_ORGANISER,
-  EVENTS__DETAILS_LOADED,
 } from '../../../client/actions'
 
 const initialState = {
@@ -29,11 +28,6 @@ export default (state = initialState, { type, result }) => {
       return {
         ...state,
         selectedOrganisers: result,
-      }
-    case EVENTS__DETAILS_LOADED:
-      return {
-        ...state,
-        ...result,
       }
     default:
       return state
